improved code quality a lot, implemeneted checkitem and removeitem in quicklist, and a lot of other minor things
