--- components/ShoppingListDoneItem/ShoppingListDoneItem.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import { View } from 'react-native';
-import { Button, Text, Icon } from 'react-native-elements';
-
-import styles from './ShoppingListDoneItemStyle.js';
-
-const ShoppingListDoneItem = ({ onButtonPress, item }) => {
-  return (
-    <View style={styles.shoppingListItemWrapper}>
-      <Text style={styles.itemName}>{item.name}</Text>
-      <Button
-        onPress={() => onButtonPress(item)}
-        icon={<Icon name="remove" type="font-awesome" size={20} color="white" />}
-        buttonStyle={styles.removeButton}
-      />
-    </View>
-  );
-};
-
-export default ShoppingListDoneItem;
