--- components/index.js
@@ -1,7 +1,0 @@
-import Navbar from "./Navbar/Navbar.js";
-import PageContainer from "./PageContainer/PageContainer.js";
-import DrawerContent from "./DrawerContent/DrawerContent.js";
-import ShoppingListItem from './ShoppingListItem/ShoppingListItem.js';
-import ShoppingListDoneItem from './ShoppingListDoneItem/ShoppingListDoneItem.js';
-
-export { Navbar, PageContainer, DrawerContent, ShoppingListItem, ShoppingListDoneItem };
