--- misc/helpers.js
@@ -1,38 +1,0 @@
-import { Keyboard } from 'react-native';
-import Clipboard from 'expo-clipboard';
-
-export const guidGenerator = () => {
-  var S4 = function () {
-    return (((1 + Math.random()) * 0x10000) | 0).toString(16).substring(1);
-  };
-  return S4() + S4() + '-' + S4();
-};
-
-export const getWeekDay = () => {
-  var d = new Date();
-  var n = d.getDay();
-  switch (n) {
-    case 0:
-      return 'Sunday';
-    case 1:
-      return 'Monday';
-    case 2:
-      return 'Tuesday';
-    case 3:
-      return 'Wednesday';
-    case 4:
-      return 'Thursday';
-    case 5:
-      return 'Friday';
-    case 6:
-      return 'Saturday';
-    default:
-      return '';
-  }
-};
-
-export const copyToClipboard = async (string, setCopiedStringTimeCallback) => {
-  Clipboard.setString(string);
-  const text = await Clipboard.getStringAsync();
-  setCopiedStringTimeCallback(text);
-};
