--- misc/variables.js
@@ -1,15 +1,0 @@
-// lengths
-export const MAX_LENGTH_ITEM_INPUT = 36;
-export const MAX_LENGTH_SHOPPING_LIST_NAME = 24;
-
-// colors
-export const APP_PRIMARY_COLOR = '#FA871A';
-
-export const DARK_RED = '#D94242';
-export const LIGHT_RED = '#FF0000';
-
-export const GREY = '#9F9F9F';
-export const DARK_GREY = '#5D5D5D';
-export const LIGHT_GREY = '#D8D8D8';
-
-export const LIGHT_GREEN = '#50D942';
