+++ src/components/index.js
@@ -1,0 +1,8 @@
+import Navbar from './Navbar/Navbar.js';
+import PageContainer from './PageContainer/PageContainer.js';
+import DrawerContent from './DrawerContent/DrawerContent.js';
+import ShoppingListItem from './ShoppingListItem/ShoppingListItem.js';
+import ShoppingListDoneItem from './ShoppingListDoneItem/ShoppingListDoneItem.js';
+import ListPage from './ListPage/ListPage.js';
+
+export { Navbar, PageContainer, DrawerContent, ShoppingListItem, ShoppingListDoneItem, ListPage };
